define: memo-databaseファイルを作成、MemoDatabaseを移行
